+++ go/idx-template.json
@@ -16,7 +16,7 @@
           "web": "Web server",
           "web-template": "Web server with HTML templates",
           "api": "API Server",
-          "api-GorillaMux": "API-GorillaMux"
+          "api-gingonic": "API-Gingonic"
         }
       }
     ]
